LPS-63096 Rename com.liferay.portal.wab.extender -> com.liferay.portal.osgi.web.wab.extender
