+++ modules/portal-osgi-web/portal-osgi-web-wab-extender/bnd.bnd
@@ -1,7 +1,7 @@
 Bundle-Name: Liferay Portal OSGI WEB WAB Extender
-Bundle-SymbolicName: com.liferay.portal.wab.extender
+Bundle-SymbolicName: com.liferay.portal.osgi.web.wab.extender
 Bundle-Version: 1.0.0
-Export-Package: com.liferay.portal.wab.extender.ws.axis
+Export-Package: com.liferay.portal.osgi.web.wab.extender.ws.axis
 Import-Package:\
 	javax.servlet;version='[2.3.0,4.0.0)',\
 	javax.servlet.http;version='[2.3.0,4.0.0)',\
