+++ modules/portal-osgi-web/portal-osgi-web-wab-extender/src/main/java/com/liferay/portal/osgi/web/wab/extender/internal/adapter/ServletExceptionAdapter.java
@@ -1,0 +1,74 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.portal.osgi.web.wab.extender.internal.adapter;
+
+import java.io.IOException;
+
+import javax.servlet.Servlet;
+import javax.servlet.ServletConfig;
+import javax.servlet.ServletException;
+import javax.servlet.ServletRequest;
+import javax.servlet.ServletResponse;
+
+/**
+ * @author Raymond Augé
+ */
+public class ServletExceptionAdapter implements Servlet {
+
+	public ServletExceptionAdapter(Servlet servlet) {
+		_servlet = servlet;
+	}
+
+	@Override
+	public void destroy() {
+		_servlet.destroy();
+	}
+
+	public Exception getException() {
+		return _exception;
+	}
+
+	@Override
+	public ServletConfig getServletConfig() {
+		return _servlet.getServletConfig();
+	}
+
+	@Override
+	public String getServletInfo() {
+		return _servlet.getServletInfo();
+	}
+
+	@Override
+	public void init(final ServletConfig servletConfig) {
+		try {
+			_servlet.init(servletConfig);
+		}
+		catch (Exception e) {
+			_exception = e;
+		}
+	}
+
+	@Override
+	public void service(
+			ServletRequest servletRequest, ServletResponse servletResponse)
+		throws IOException, ServletException {
+
+		_servlet.service(servletRequest, servletResponse);
+	}
+
+	private Exception _exception;
+	private final Servlet _servlet;
+
+}
