--- modules/portal-osgi-web/portal-osgi-web-wab-extender/src/main/java/com/liferay/portal/wab/extender/internal/WabUtil.java
@@ -1,45 +1,0 @@
-/**
- * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
- *
- * This library is free software; you can redistribute it and/or modify it under
- * the terms of the GNU Lesser General Public License as published by the Free
- * Software Foundation; either version 2.1 of the License, or (at your option)
- * any later version.
- *
- * This library is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
- * details.
- */
-
-package com.liferay.portal.wab.extender.internal;
-
-import java.util.Dictionary;
-
-import org.osgi.framework.Bundle;
-import org.osgi.framework.Constants;
-
-/**
- * @author Raymond Augé
- */
-public class WabUtil {
-
-	public static String getWebContextPath(Bundle bundle) {
-		Dictionary<String, String> headers = bundle.getHeaders();
-
-		return headers.get("Web-ContextPath");
-	}
-
-	public static boolean isFragmentBundle(Bundle bundle) {
-		Dictionary<String, String> headers = bundle.getHeaders();
-
-		String fragmentHost = headers.get(Constants.FRAGMENT_HOST);
-
-		if (fragmentHost == null) {
-			return false;
-		}
-
-		return true;
-	}
-
-}
