--- modules/portal-osgi-web/portal-osgi-web-wab-extender/src/main/java/com/liferay/portal/wab/extender/internal/configuration/WabExtenderConfiguration.java
@@ -1,34 +1,0 @@
-/**
- * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
- *
- * This library is free software; you can redistribute it and/or modify it under
- * the terms of the GNU Lesser General Public License as published by the Free
- * Software Foundation; either version 2.1 of the License, or (at your option)
- * any later version.
- *
- * This library is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
- * details.
- */
-
-package com.liferay.portal.wab.extender.internal.configuration;
-
-import aQute.bnd.annotation.metatype.Meta;
-
-import com.liferay.portal.configuration.metatype.annotations.ExtendedObjectClassDefinition;
-
-/**
- * @author Iván Zaera
- */
-@ExtendedObjectClassDefinition(category = "platform")
-@Meta.OCD(
-	id = "com.liferay.portal.wab.extender.internal.configuration.WabExtenderConfiguration",
-	localization = "content/Language", name = "%wab.extender.configuration.name"
-)
-public interface WabExtenderConfiguration {
-
-	@Meta.AD(deflt = "60000", required = false)
-	public long stopTimeout();
-
-}
