+++ modules/portal-osgi-web/portal-osgi-web-wab-extender/src/test/java/com/liferay/portal/osgi/web/wab/extender/internal/definition/MockServletContextListener.java
@@ -1,0 +1,33 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.portal.osgi.web.wab.extender.internal.definition;
+
+import javax.servlet.ServletContextEvent;
+import javax.servlet.ServletContextListener;
+
+/**
+ * @author Miguel Pastor
+ */
+public class MockServletContextListener implements ServletContextListener {
+
+	@Override
+	public void contextDestroyed(ServletContextEvent servletContextEvent) {
+	}
+
+	@Override
+	public void contextInitialized(ServletContextEvent servletContextEvent) {
+	}
+
+}
