+++ modules/portal-osgi-web/portal-osgi-web-wab-extender/src/test/java/com/liferay/portal/osgi/web/wab/extender/internal/definition/WebXMLDefinitionLoaderTest.java
@@ -1,0 +1,114 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.portal.osgi.web.wab.extender.internal.definition;
+
+import java.net.URL;
+
+import java.util.EventListener;
+import java.util.List;
+import java.util.Map;
+
+import javax.servlet.Servlet;
+import javax.servlet.ServletContextListener;
+
+import javax.xml.parsers.SAXParserFactory;
+
+import org.apache.felix.utils.log.Logger;
+
+import org.junit.Assert;
+import org.junit.Test;
+import org.junit.runner.RunWith;
+
+import org.mockito.Mock;
+
+import org.osgi.framework.Bundle;
+
+import org.powermock.modules.junit4.PowerMockRunner;
+
+/**
+ * @author Miguel Pastor
+ */
+@RunWith(PowerMockRunner.class)
+public class WebXMLDefinitionLoaderTest {
+
+	@Test
+	public void testLoadCustomDependencies() throws Exception {
+		Bundle bundle = new EntryLoaderMockBundle();
+
+		WebXMLDefinitionLoader webXMLDefinitionLoader =
+			new WebXMLDefinitionLoader(
+				bundle, SAXParserFactory.newInstance(), new Logger(null));
+
+		testLoadDependencies(webXMLDefinitionLoader, 1, 1, 1);
+	}
+
+	@Test
+	public void testLoadDefaultDependencies() throws Exception {
+		Bundle bundle = new MockBundle();
+
+		WebXMLDefinitionLoader webXMLDefinitionLoader =
+			new WebXMLDefinitionLoader(
+				bundle, SAXParserFactory.newInstance(), new Logger(null));
+
+		testLoadDependencies(webXMLDefinitionLoader, 0, 0, 0);
+	}
+
+	protected void testLoadDependencies(
+			WebXMLDefinitionLoader webXMLDefinitionLoader, int numfOfListeners,
+			int numOfFilters, int numfOfServlets)
+		throws Exception {
+
+		WebXMLDefinition webXMLDefinition = webXMLDefinitionLoader.loadWebXML();
+
+		List<ListenerDefinition> listenerDefinitions =
+			webXMLDefinition.getListenerDefinitions();
+
+		Assert.assertEquals(numfOfListeners, listenerDefinitions.size());
+
+		for (ListenerDefinition listenerDefinition : listenerDefinitions) {
+			EventListener eventListener = listenerDefinition.getEventListener();
+
+			Assert.assertTrue(eventListener instanceof ServletContextListener);
+		}
+
+		Map<String, ServletDefinition> servletDefinitions =
+			webXMLDefinition.getServletDefinitions();
+
+		Assert.assertEquals(numfOfServlets, servletDefinitions.size());
+
+		Map<String, FilterDefinition> filterDefinitions =
+			webXMLDefinition.getFilterDefinitions();
+
+		Assert.assertEquals(numOfFilters, filterDefinitions.size());
+	}
+
+	@Mock
+	private Servlet _servlet;
+
+	@Mock
+	private ServletContextListener _servletContextListener;
+
+	private static class EntryLoaderMockBundle extends MockBundle {
+
+		@Override
+		public URL getEntry(String path) {
+			Class<?> clazz = getClass();
+
+			return clazz.getResource("dependencies/custom-web.xml");
+		}
+
+	}
+
+}
