+++ modules/portal-osgi-web/portal-osgi-web-wab-extender/src/test/resources/com/liferay/portal/osgi/web/wab/extender/internal/definition/dependencies/custom-web.xml
@@ -1,0 +1,30 @@
+<?xml version="1.0"?>
+
+<web-app
+		metadata-complete="true"
+		version="2.6"
+		xmlns="http://java.sun.com/xml/ns/javaee"
+		xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+		xsi:schemaLocation="http://java.sun.com/xml/ns/javaee http://java.sun.com/xml/ns/javaee/web-app_2_6.xsd"
+>
+	<listener>
+		<listener-class>com.liferay.portal.osgi.web.wab.extender.internal.definition.MockServletContextListener</listener-class>
+	</listener>
+	<filter>
+		<filter-name>Mock Filter</filter-name>
+		<filter-class>com.liferay.portal.osgi.web.wab.extender.internal.definition.MockFilter</filter-class>
+	</filter>
+	<filter-mapping>
+		<filter-name>Mock Filter</filter-name>
+		<url-pattern>/mock_filter</url-pattern>
+	</filter-mapping>
+	<servlet>
+		<servlet-name>Mock Servlet</servlet-name>
+		<servlet-class>com.liferay.portal.osgi.web.wab.extender.internal.definition.MockServlet</servlet-class>
+		<load-on-startup>0</load-on-startup>
+	</servlet>
+	<servlet-mapping>
+		<servlet-name>Mock Servlet</servlet-name>
+		<url-pattern>/mock_servlet</url-pattern>
+	</servlet-mapping>
+</web-app>
