+++ source-formatter.properties
@@ -200,7 +200,7 @@
         **/markdown-converter/**,\
         **/osgi-util/**,\
         **/portal-tools-shard-builder/**,\
-        **/portal-wab-extender/**,\
+        **/portal-osgi-web-wab-extender/**,\
         **/registry-api/**,\
         **/sync-engine/**,\
         modules/test/poshi-runner/src/main/java/com/liferay/poshi/runner/util/GetterUtil.java,\
@@ -233,7 +233,7 @@
     #
     secure.xml.excludes=\
         modules/apps/xsl-content/xsl-content-web/src/main/java/com/liferay/xsl/content/web/util/XSLContentUtil.java,\
-        modules/portal-osgi-web/portal-osgi-web-wab-extender/src/main/java/com/liferay/portal/wab/extender/internal/definition/WebXMLDefinitionLoader.java,\
+        modules/portal-osgi-web/portal-osgi-web-wab-extender/src/main/java/com/liferay/portal/osgi/web/wab/extender/internal/definition/WebXMLDefinitionLoader.java,\
         modules/sdk/gradle-plugins-maven-plugin-builder/src/main/java/com/liferay/gradle/plugins/maven/plugin/builder/util/XMLUtil.java,\
         portal-impl/src/com/liferay/portal/security/xml/SecureXMLFactoryProviderImpl.java,\
         portal-impl/src/com/liferay/portal/tools/JavadocBuilder.java,\
